Add generic query/execute functions
This change removes read/write specific query and execute functions from
the upstream connectors. These read/write specific functions relied on
the ability of the adapter to parse the query to determine whether it
would return a result with rows. If the adapter was unable to parse a
query, it could not return the correct response to the user in all
cases.

The specific read/write functions were replaced with generic query() and
execute() functions. These functions parse the resultset returned by
upstream to determine what should be sent to the user.

This change also removes read/write specific upstream prepared
statements, as the execute() function will now accept any prepared
statement.

Change-Id: If1e54fcbc1781f0c9016d23a1680dbd2e054bfeb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/424
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>