nom-sql: `groups` is a reserved keyword now
Per https://dev.mysql.com/doc/refman/8.0/en/keywords.html, "GROUPS" is a
reserved keyword in mysql as of 8.0.2, so we need to add it to our
reserved keywords list so that when formatting a column or table named
`groups` we properly escape it.

As a side note, this way of representing reserved keywords is majorly
icky, as evidenced by having to shift things down due to the tuple
length limit for `alt` - I'm going to work on a more robust approach at
some point in the future.

Change-Id: I9895a6383c1102cab8b4e69586920872da5043c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/410
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>