Convert race reads to mirror reads
Previously we tried to run any read query against both upstream and
noria in separate tasks if the race reads feature was enabled. We've
decided that this feature should instead become mirror reads.

Starting a new connection to satisfy the ownership requirements of tokio
task when running the query against the upstream database meant that we
broke prepared statements against upstream, since those are always
prepared over a given connection.

Since this feature is designed (for now) to be purely about data
collection using the query coverage info setup, we have decided to
change this to act in a serial fashion, and moreover to always return
the upstream database result.

This will allow us to collect the correct sequence of queries for any
given test suite or application being run against ReadySet, as it
ensures we always return the same result the upstream database would
have returned during data collection.

Change-Id: I6ae22340bea498bd02b81c08cb321547e198ae16
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/413
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>