Accept `NUMERIC` on `make_field_description`.
The `NUMERIC` type is not being handled by our PostgreSQL frontend/backend
protocol, which we need to handle (at the very least, to support AVG aggregates
in our queries, since PostgreSQL sets `NUMERIC` as the type of those columns).
For now, this fix only makes `NUMERIC` take the same type length as `FLOAT8`,
because `NUMERIC` is a floating-point number representation (that can take
`precision` and `scale` parameters), it makes sense for now to map it to our
`Double` DataType counterpart, at least for now.

Refs: ENG-568
Change-Id: I114371bb73b2e739358cffbc0725edf992317f32
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/423
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>