data: Allow coercing all int types to bool
Add support to coerce_to for coercing all integer types to booleans, by
checking if they're not 0 then delegating to <DataType as
From<bool>> (which itself just makes a u32 that's either 0 or 1
depending on the boolean, but may change eventually).

This isn't *all* of the coercion rules for boolean - specifically
postgres lets booleans be a case-insensitive prefix of the strings
'true' or 'false' - but it's a start.

Change-Id: Ib9fb79771835a94a86dc0b76f60a33f19bbfe8fb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/431
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>