noria-client: Log errors that cause fallback
If we get an error from noria that would cause a query to go to
fallback, log that error at the ERROR log level with tracing - whether
we have fallback configured or not. Also, log at the WARN level if a
query goes to fallback because we're in a transaction.

Change-Id: If2c6a33ecc158c154c2c89606b9d3464102680ad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/434
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>