psql-srv: Add BackendMessage::ParameterStatus
Add the ParameterStatus postgresql backend message to our BackendMessage
enum, and implement serialization for it to the client. This is
used (among other things) for sending parameters about the connection to
the client as part of the startup flow, and is necessary now so we can
send client_encoding, TimeZone, and DateStyle to the psycopg client.

Change-Id: Icb26d4e0ebd4a6bef47ea02ff7072fd7c0d41d41
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/437
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>