psql-srv: Allow any number of messages in a response
Previously nothing in the psql-srv protocol implementation necessitated
ever sending more than 2 messages to the client at once - so to avoid
the extra heap reference+allocation we just had Response::Message and
Response::Message2 variants - now, we'd like to be able to send more
than 2 messages (in the particular case, we'd like to send multiple
ParameterStatus messages after AuthenticationOk). The optimization of
avoiding the heap-reference is still sound, though - fortunately,
there's a crate that gives us the notion of "use the stack up to n
elements, use the heap post-n" called smallvec we can use here. I picked
2 as the value of n to use here since that matches Message2, but there
may be more science worth doing there in the future once we start
fine-grained optimization of this crate.

Change-Id: I0e8c47e6b7a843595538f42a41fd986c510dcb5a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/438
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>