logictests/irl.test: Setup for execution timing
Set up the irl logictest for simple benchmarking execution timing with
the new --time logictest flag, by duplicating queries first before all
writes (to time migrations), then after the first upquery (to time warm
reads).

Empirically, some queries are slow the first two times they're run
rather than just the first time, then the third time seems to be full
cache warmup, so these queries have a hidden version in between the
first and the third to make sure we're as fast as possible before we
record warm execution time. I'm not sure *why* this is happening, and it
might be worth looking into in the future, but for now it's probably not
a huge deal.

Change-Id: I5dcd66e8fada09b571328f9ecee5776382092671
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/447
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>