data: Allow FLOAT{4,8} in ToSql impl for DataType
DataType can be converted to floats and doubles - there's even a pattern
match arm for them - but the types need to be in the `accepts!` to
prevent to_sql_checked from returning an error. This is all that's
necessary to get both real and double working round-trip through
noria-psql, so this commit also de-ignores those two type tests.

Fixes: ENG-547, ENG-548
Change-Id: I8fddd60eec1abdc04acf92b7764640c4500fc207
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/449
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>