Add fail_no_upstream conditional to logictests
This conditional will expect a statement to fail if no replication url
is given (i.e., we are operating against noria only). The query will be
considered pass if it fails, and a fail if it passes.

This allows us to set logictests that fail against noria but pass
against noria + replication as passing tests. Any queries that fail
against noria only will be marked with the fail_no_upstream conditional.
When these statements do pass against noria, the logictest will report a
failure.

The alternative to this approach would have been to invert the results
on tests that do use a replication url. However, this approach is not
ideal, as it may obscure unexpected failures that would occur later in
the test.

In this approach, all queries that fail on noria are marked as such, and
it would be obvious if a regression causes a later query to fail.

Change-Id: I066bd0229e047927213c8ac7d63452ca00acc625
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/452
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>