replicators: Ignore pgsql_replication test for now
This test is failing a bunch in CI but can't be run locally (nor does it
seem like it could *ever* have been run locally) due to an issue where
the runtime int he TestHandle gets dropped while the runtime for the
test itself is still active, giving the error message "Cannot drop a
runtime in a context where blocking is not allowed. This happens when a
runtime is dropped from within an asynchronous context.". Without the
ability to debug it locally it's going to be difficult to get it
consistently passing again, so I'm just ignoring it until the original
author can help debug the test. We're testing postgres replication
transiently via other tests that do fallback, so I'm not super worried
that we're going to accidentally break replication while this is
disabled.

Change-Id: I09f1c5221963c913265ff07c11afb456748399c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/454
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI