mir: Make MIR graphviz lazy
Currently, MIR's to_graphviz function allocates and return a string
directly. This is costly on its own since it recursively calls into
other nodes which might perform a lot of unused separate heap
allocations - but is especially bad in the case of logging where we
might construct that string only to discard it immediately because the
log event is below the current (dynamic) log level filter. This commit
splits out the methods of the trait into a (new) `graphviz_fmt` method
which rather than *returning* a string takes a `fmt::Formatter` (a'la
Display or Debug) which it writes to, and then makes the existing
`to_graphviz` methohd return a `GraphVizzed` wrapper struct around
`self`, whose `Display` impl delegates to `graphviz_fmt`. This
essentially makes it so that simply calling `to_graphviz()` on a MIR
query does no work and performs no allocations, until the return value
is used as part of a `format!()` call - which makes it so that we can
filter out log events with graphviz without doing any throwaway work.

Change-Id: Ib4ec1e7df838824b2a7ce59fc40ed68e68e26551
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/462
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI