Add crate with definitions for buildkite pipelines
Add a new crate with Rust data structures representing the structure of
a Buildkite pipeline, and serde Serialize and Deserialize impls that
allow reading and writing those pipelines to and from buildkite pipeline
yaml files. This includes some unit and property tests for specific bits
of pipelines, plus an integration test suite that tests parsing all
of *our* pipelines specifically.

The hope is that we can migrate all of the dynamic pipeline generation
we do to use these data structures, rather than textually templating
YAML, and get the advantage of some nice type safety and rust data
structure manipulation.

Change-Id: I38c2aa8f2756b006ed421b954f96ee1f0205b34d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/468
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>