Handle Postgres response.
This commit adds the logic to handle an upstream db response to a query.
The logic is straight forward: we take the Postgres `Row` type and we extract
the name and type of the columns, and we create a basic `Resultset` with them,
projecting all columns (since this is not coming from Noria, all columns that
are present in the `Row`s are columns that are meant to reach the user).

The `FromSql` trait was implemented for `DataType` in order to retrieve the
values from the Postgres `Row` type.

Refs: ENG-584
Change-Id: I8bad623a02f80d9321d894d7b76c561fa0cb212b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/475
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>