mir: configurably return unsupported for topk
Since we're going live with customers soon and we still crash and burn
if a topk node ever gets a negative delta that brings it below k
records, we want to make queries with topk go through fallback. We still
have tests that use topk, though, so this behavior has to be
configurable. This commit introduces a new `sql::mir::Config` struct,
with an `allow_topk` field that defaults to `false`, threads that value
all the way down from the top-level clap opts, through
ControllerState, Recipe, and SqlIncorporator to SqlToMirConverter, and
then sets the flag to `true` in all places we want topk to be
allowed (integration tests and logictests).

This commit can likely be reverted wholesale once we properly support
TopK

Fixes: ENG-588
Change-Id: I359b7dfbc5e88ce1ae378e185beca93ea2a5a0f7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/478
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>