ops/cfn: Beginnings of a AWS Cloudformation stack
This will be consumed by our customers and it appears to be the way that
AWS encourages people to distribute their software. This was not
terribly hard to write to get this far and we can more easily make use
of a number of tools for exactly this usecase that AWS offers us.

Change-Id: I435b71f2a25df23755b34635a841cf57dc1ac70a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/479
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>