Change shim return types to io::Error
Previously we had a restriction that Self::Error for an implementor of
the MysqlShim had to implement From<io::Error>. There is even a doc
comment explaining that this is to allow us to bubble io::Errors up the
chain resulting in conn drops (on purpose).

Unfortunately in practice this is not how this ended up getting used.
Instead we lazily just returned an owned error enum, of which only one
variant could be derived from io::Error. To top it off, io errors would
sometimes be embedded in other error enum types, such as mysql async
errors as one example.

This commit changes the return types to very explicitly be io::Error,
forcing on_query, on_execute etc. to have to deal with writing back
appropriate error packets, or bubble up io errors explicitly if they
are encountered.

This also ensures that we dig out io errors out of the various error
enum variants they may be inside of.

This commit also explicitly writes an io error to drop our connection
to the client in cases where we lost our connection with the setup
fallback database.

This is necessary because connections have a 1:1:1 relationship between
client:adapter:mysql. If we lose our connection to mysql we need the
client to re-initiate their connection to the adapter so that we attach
that to a new connection to mysql. Any ORM or database driver at this
point would then send all their SET statements back down the connection
(as well as not assume previous prepares are still prepared on that
connection).

Change-Id: I6314891df6e2f8b75e0f8f0f3b5b9bae3db29f18
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/480
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>