Already applied fixes to get new buildkite stack
The stack name was not valid as it existed, so replaced that and
improved the local validation.

Cloudformation was hanging on the latest version of the stack so
downgraded one version until we have more time to figure out why.

Policies on secrets do not work unless the role is there so updating
policies outside of Terraform for now.

Typo with the name of the secret parameter.