nom-sql: Parse arbitrary expressions in LIMIT/OFFSET
SQL supports any constant-valued expression in the LIMIT and OFFSET
clause of a query, but previously we only parsed unsigned integers. We
aren't going to *support* arbitrary expressions here yet, but there are
cases where we can statically determine that a query can only ever
return one result and so we can safely strip out a limit clause, and
we'd like to be able to do this if the limit clause is any
expression (in this case, it's a placeholder).

This commit updates the LimitClause AST struct to use any expression for
both limit and offset, changing offset to an Option rather than
defaulting it to 0 in the parser, and then updates all callsites to
handle arbitrary expressions - returning an unsupported error in
sql->mir if the limit expression is anything other than an integer
literal.

In order to get this parsing happily, since this is the first case of an
expression at the *end* of the input provided to a select statement,
this required wrapping some stuff in the expression parser in nom
`complete!` macros.

Fixes: ENG-601
Change-Id: I59e225d89b9b32c3aa1974750f8a267ed547be77
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/483
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>