Buildkite should not exclude steps if there is an ops file modified.
The current pipeline excludes certain steps if an ops file is
modified based on the `exclude` option in diff-conditional. However,
it is possible CL's include both ops and code changes, for example
changing a command line argument that needs to be propagated to
new build images. To fix this, we should only include steps if a
related file is modified, rather than exclude stpes if a file is
modified, allowing both the ops pipeline and dev pipeline to
run.

Change-Id: Ib5462244720e7377215adfe43979861da0c577fb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/489
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>