shell.nix: Bump nixpkgs pin
Bump the version of nixpkgs referenced in the shell.nix.

I'm still the only person who's using this, so at some point it may end
up just making more sense to remove and have me gitignore it, but for
the time being I'd like to keep it up to date.

Change-Id: I46d4c8319da14688f29dad47494dd29e5e04cdb0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/495
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>