ci: Still build stuff if Cargo.{lock,toml} change
The glob that's here for all of the Rust-related build stuff appears to
only applies to children of subdirectories that aren't ops, not
top-level files. There are a couple of top-level files that changes to
would want to trigger a rebuild, specifically the cargo config and
lockfile, and the rust-toolchain file.

Change-Id: I90a1a73c54f988da4c868f44678a17681868280f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/497
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>