Create way to write error from row writer
Currently once we create a row writer we consume our query result
writer, and no longer have a way to write error packets back. This is
problematic because we often need to write back error packets as we
discover issues in the middle of writing out rows.

This creates a method so that can call down to the query result writer
we have taken ownership of in the row writer so we can successfully
write an error packet back to the caller.

Change-Id: I9067c9b80e644b3e7a4c518e576fc1364e62bd75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/503
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>