Changed nom-sql's TableKey::PrimaryKey to include a name field
This is catering to undocumented behavior in MySQL:  Apparently, when
you ALTER TABLE ... ADD PRIMARY KEY or CREATE TABLE ... (..., PRIMARY
KEY), you're allowed to name the key like any other index.  This name
does not appear to be saved anywhere, but the syntax is accepted by the
database.

And, naturally, Flarum issues ALTER TABLEs that look like this.

Change-Id: I333506b798edaa4d34719416ed157964c4cbc2ca
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/534
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI