Add order_limit_removal pass
Added a pass to remove the limit and order fields of a query when it can
be determined that the query will return a single result.

This determination is made by checking whether the query includes an
equality conditional comparing a primary or unique key to a literal.
This pass could be expanded to catch other conditionals that would
result in at most one row returned (e.g., a conditional compary a unique
key to a select statement that returns a single result).

This pass requires that all columns have explicit table names, and must
be run after the expand_implied_tables() pass.

Change-Id: I7e46ea87f16a86fa0804b6f92b95672964907d2c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/506
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>