nom-sql/src/compound_select.rs:  Add ignored test for an insane compound SELECT
Naturally, this is issued by Flarum.  I don't currently have a fix for
this; it fails near the end, parsing "order by is_sticky and not exists
(...)"; the ORDER BY clause parses, but it bombs when it hits "and not
exists".

It smells like we're expecting parentheses and not getting them, but
unconfirmed.

Change-Id: I7d471ad8dcb0478fedbcf85647067baf47816848
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/538
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>