Logictest support for any authority and fix for consul+logictest.
This allows the logictest `Verify` subcommand to specify an
authority: local, consul, or zookeeper. This allows testing
end-to-end authority behavior for consul and zookeeper with
logictests, previously only local authority was used. The
logictest framework now uses the script name as the
deployment name to keep keys isolated in the authority.

Along with this is several consul authority fixes that
result from repeating multiple tests with the same
deployment name.

The default authority is set by the ci docker-compose:
build/docker-compose.ci-test.yaml. As a result, this
CL changes the authority to zookeeper.

Change-Id: I705ec06c77c5fd9793b7960591bb3af376aa90c1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/517
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>