nom-sql:  Add support for CHANGE/MODIFY in ALTER TABLE statements
These are used _extensively_ in the Flarum migrations.

It was necessary to add CHANGE and MODIFY as new known keywords; other
than that, this is a straightforward addition of some new parsing logic.

Semantically, CHANGE and MODIFY are the same, with the only difference
being that CHANGE can change the name of a column (and requires you to
repeat the name if you wish not to), so they both parse down to CHANGE
internally.  If round-tripped, a MODIFY will turn into a CHANGE.

Change-Id: Ia7c601cfa6569d45eafb7f599ff583b6cb1d2a21
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/526
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>