ops/aws-accounts: Integrate with buildkite pipeline
Integrate the pipeline for building the buildkite queue uploader and
running it to make buildkite steps for all of the terraform directories
that we need to run validate and plan on with the overall monorepo
pipeline. The steps that were here to run cargo clippy and cargo test
and stuff are already taken care of by the existing pipeline, so those
no longer need to exist, nor does the dockerfile for the docker image
that we were running them in.

Change-Id: Ifde2e30b440b05fb2816382c66bc123b4d4f043a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/542
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>