Run clustertests in the nightly CI pipeline.
The clustertest framework mimics a real-world deployment with
a multi-process deployment. This allows us to test systems
properties such as failure behavior and recovery, leader
election, replication, etc.

This CL brings it into CI in the nightly pipeline. Depending
on the performance in CI it may be moved to the per-commit
pipeline.

This CL is accompanied by several clustertest changes that
enable it to work in CI. These changes are:
 - Removing building and launching docker containers, CI
   already has this.
 - Support for using any arbitrary authority (such as consul).

Change-Id: I8674828395fcbe64795ee33893fb2c75d2f0d9dc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/545
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>