ops: Auto-configure buildkite secret policy
Previously, every time we set up a new buildkite queue we'd have to
manually go in to the agent token secretsmanager secret and add the
necessary IAM roles created by the cloudformation stack (the instance
role, and the role used by the autoscaling lambda to execute) to the
list of roles allowed to access the secret. Getting this to happen
automatically is a little annoying, since buildkite's cloudformation
stack doesn't include the ARNs of the relevant roles as outputs. To
complicate that even further, the *names* of the roles aren't even
entirely consistent - if the name of the stack itself is too long, the
"AutoscalingLambdaExecutionRole" suffix gets chopped off, eg in
"buildkite-c5a-4xlarge-AutoscalingLambdaExecutionRo-1BTIMEL4G04JE".
Given all of this, the most reliable thing we've got is to create a data
source that's explicitly dependent on the aws_cloudformation_stack
resource for the buildkite queue itself that just looks up *all* roles
starting with the name of the stack, and then thread the resulting ARNS
through the buildkite_queue_shared module to a new
`aws_secretsmanager_policy` resource granting all those roles access to
the secret. This slightly over-provisions, since only the instance role
and the autoscaling lambda execution role actually need access and this
also gives access to the AsgProcessSuspenderRole, but I figured since we
can't *really* robustly filter on the autoscaling lambda role, and since
the buildkite agent token is a relatively low-impact secret, it wouldn't
be a huge deal to grant slightly more permissions there.

Change-Id: I82ee842ced3e46f0e27d8d9a1f1aa65ae3bf9063
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/546
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>