Unsupport drop/alter table in prepare
I missed this in a previous CL where we decided to catch all the rest of
these to just go to upstream. Unfortunately we can't send drop or alter
table to upstream currently because we shouldn't support executing
these. Unfortunately we have no way of mapping back to the original
prepared statement during execution (we just go by statement id at that
point), so we also have no way of conditionally supporting different
query types during execution. Therefore we have to block them during the
prepare stage if they are truly unsupported.

Change-Id: Ifb3f5ab827252c81db7e78b03724e500acc080f3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/547
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>