new-images/consul-server: Add user data init script
Add a user-data-init.sh script to the consul-server image in new-images,
which sets up the data volume for /opt/consul and templates the
/etc/consul.d/consul.hcl file with runtime-configured values like the
bootstrap_expect value and the retry_join tags. There's still
a *default* consul.hcl config file that we put there as part of the
install - I figured I'd leave it there so the server can start up
without failing, but it might be better to remove it instead.

Change-Id: I5d913be8ccb8b7c3ec564885b5f9d368f0577f5f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/551
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>