Added support for BYTEA type.
A new `DataType::ByteArray` was added, which holds a `Vec<u8>`.
Couple of notes to make here:
1. The `nom-sql` support to parse a literal byte arrays is a naive one, and we
are only parsing hex-formatted bytea literals (see BYTEA documentation).
2. We are now able to transform _any_ byte array into a DataType: if it cannot
be transformed into one of the Text variants, then it's transformed as a
`DataType::ByteArray`.
3. We are by default displaying the byte array into hex format (as this is the
default in PostgreSQL). However, it's possible to configure that in PostgreSQL,
and an alternative to that is using the escaped format. We should add support
for this in the future.

Refs: ENG-600
Change-Id: I82c7b0061f08c5699e04fc3aa0aa04d82c03f3ec
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/555
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>