ops/build: Import readyset pipeline into terraform
Add the buildkite terraform provider to the build/global module, and add
a terraform definition for the "readyset" buildkite pipeline (our
default monorepo pipeline), which has been already imported into the
state via terraform import and confirmed to be functionally
unchanged (there's some slight YAML indentation changes, but it doesn't
make a difference with how the pipeline works).

In order to get this working with plan and apply (including in CI), I've
added services@readyset.io to buildkite and generated an API token for
it, which is both in 1password and in the buildkite secrets for the ops
queue. To run things locally, define the BUILDKITE_API_TOKEN env var to
be this token.

As we add new buildkite pipelines over time, it'll be nice to manage
them all in terraform so that the configuration is tracked and
declarative.

Change-Id: I37b5b0562ba842183df3e562b70446968b38a3e8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/556
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>