ops/new-images: Promote to ops/images
Remove ops/images, and promote ops/new-images to ops/images, removing
the old ops packer build steps and updating all references in CI from
new-images to images.

We haven't touched the old images in long enough and we're focusing
enough of our effort on the new images now that I figure it's time to
just get rid of the old ones.

Change-Id: Ic7937872008eba656257ff13f0b6b55eede93317
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/559
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>