ops/cfn: Consul template in super template
This creates a new template that handles setting up a Consul cluster and
then then adds it to the super template that contains everything else.

This also sets up an RDS instance as part of the super template so
everything required for a working instance is created with a single
stack.

Change-Id: I2091ceb2ce55fc86fa125a8683853269b7fe034f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/560
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>