images: Add consul-client to images
Add a consul-client to both the readyset-mysql-adapter and
readyset-server images, using a version of the existing file+script
provisioner for consul client updated to also configure at runtime from
the user-data-init.sh script in the same way that the consul server
does.

This is untested as of this commit, but hopefully that'll change soon!

Change-Id: Ic67c46e97c21c2e8d9dabad630b55b37c562b532
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/568
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>