ops/image-deploy: Packer config to copy images
AMIs cannot be copied between accounts and regions without starting an
instance and then turning it off again. The best tool for that at the
moment is Packer.

This takes advantage of the manifest file that Packer can generate from
our image build to pass the AMI IDs between different processes.

Change-Id: I0dbe602edadb0ed5c64c317d482b6d71c743dfa3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/569
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>