substrate: Mostly upgrade to 2021.09
Ran all the required commands as well as upgraded a number of the
environments.

prod was not upgraded because it came back with an error and is
currently unused.

network was not upgraded as it would bump the Tailscale subnet routers
which are still in need of fixing.

Substrate now includes a number of symlinks which are not copied
between Docker images. So moving installing to the final Docker image.

Change-Id: I923711b3aac91ccb17d93ebcdf9a56c9003d95dd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/576
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Terraform-Plan-Review: Griffin Smith <griffin@readyset.io>