Fix types.rs tests
The main reason for the flakiness was that noria was never stopped
as it was spinned in a thread with an infinite loop and the handle
was never dropped. Moving the handle out of the loop and dropping it
after proper shutdown fixes the flakiness. In general the test flow
was messy, and is cleaner now.

It also gets rid of the ZooKeeper dependency (and the associated
Deployment). Instead it uses LocalAuthority, same as the logictests do.

In addition there was a failure in the bytea_bytes tests, as WAL seem
to prefix the hex string with an "\x", which failed to parse properly.

Fixes: ENG-620
Change-Id: I9b502c911985718cc92f8c6a0edb0f80229025b9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/588
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>