Implement Request timeouts with Tower
Previously there was no timeout implemented on requests, which was
noticed during testing of network partitioning between the adapter and
server.

Tower already has a suggested best practice for this, to use a Timeout
middleware.

This commit adds that middleware, and has been tested to fix the hanging
caused by the server initially being severed from its network.

Co-Authored-By: Peter Farr<peter@readyset.io>

Change-Id: I2cb1c94cda2c46169f1df8a52fa20e41189e47d6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/589
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>