query-generator: Allow generating parametrized IN
Add a new QueryOperation to query-generator, InParameter, which
generates queries with an IN filter where the right-hand side is a list
of placeholders. The length of the list is a parameter, but as with
other high-state-space parameters is only actually set to different
values when fuzz testing (eg in the Arbitrary impl for QueryOperation)
and is fixed to 3 when generating exhaustive tests.

Since we want to try to generate multiple different values for the
different instances of the parameter in the right-hand side of the IN,
and this is the first instance of us intentionally adding the same
column as a parameter to the query multiple times, this also introduces
a new field to the type in of QueryState::parameters, now a struct,
which is optional but if set provides an index to distinguish between
duplicate instances of the same parameter in the query. This index is
then passed to `unique_value_of_type` to get a value to pass to
`expect_value` for that column.

Refs: ENG-619
Change-Id: I09d10ebc8443301c3fd89a871786c5046852cfde
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/591
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>