logictest: Don't forget GeneratorState in Seed
When generating queries, query_generator stores metadata about those
queries, such as expected values for columns, in the GeneratorState. On
the `noria-logictest generate` side, however, we were making a
GeneratorState to generate queries for a Seed, but then *forgetting*
that GeneratorState - later, when actually running the generate process,
we'd rebuild the GeneratorState from the list of tables we'd made for
the test script. This meant that any expected values (such as, in this
case, those registered by the new InParameter query operation) wouldn't
carry through to the actual INSERT statements we'd generate for the
logictest. It's actually a relatively simple change to just add a
`generator: GeneratorState` field to the Seed struct, and write the
generator we use when generating a Seed from a Vec<QuerySeed> to that.
When generating Seeds from files, we still just use the set of tables in
the query.

Refs: ENG-619
Change-Id: Ie4f38cc43c825130638119e01948f9b87a82fefe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/594
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>