logictest: Add command for generating permutations
Add a new `permute` subcommand to noria-logictest for generating tests
from exhaustive permutations of (small) operators up to a certain depth,
with an option to limit the permutations generated to only those
including a particular operator (useful when adding new operators).

This essentially makes the process by which the current
logictests/generated/mysql directory was generated reproducible, modulo
renaming failing tests to .fail.test which may happen in the future.

Change-Id: I4704adb5d426e2a1e7334558530999e35c0f8bf5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/597
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>