logictest: Respect SortMode when generating tests
We had already been correctly setting the sort mode to RowSort on
queries that didn't have an order_by, but never actually *used* that
sort_mode when creating query records in scripts based on results
returned by the upstream database. This changes that to first sort the
results returned by the db according to the sort mode, then store the
actual sort mode on the input Query in the Query we create as part of
the output script.

Fixes: ENG-403
Change-Id: I15fb9c1be1283ed7edb92c0a8b88523a5acf55e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/602
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>