Set a timeout for http requests made by Controller
Previously we had an rpc timeout for the controller. It was quite long
(60 seconds), and also only applied if we circled back to the top of
loop ieration.

It did not apply, unfortunately, to http requests made within the
reconciliation loop.

This commit ensures that the client we create has a sane timeout set,
and we've updated the RPC timeout to a more sensible 10 seconds.

Co-Authored-By: Justin Miron<justin@readyset.io>
Change-Id: I10acbbfdeb1684dddda2fab649b2504fa268020d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/607
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>