Refactor authority thread into Controller, rename Controller*.
ControllerOuter, is an entity responsible for handling the control
plane for a noria server. This includes external requests, and
participation in leader election/liveness protocols. When the
ControllerOuter is elected leader, they create a ControllerInner
which performs all the tasks that a leader should perform.

The naming is quite confusing. To make this more clear ControllerOuter,
ControllerInner have been renamed to Controller and Leader,
respectively.

The startup of the authority thread is also moved into the
Controller as its functionality is tightly coupled with the
control plane.

Change-Id: Ifcacdcf11887df8c5b70e6ec1a3bebd50de02279
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/609
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>